lots of things
hatchrelease doesn't work but is in there
place works and is in there
no more placeCargo and placeHatch, both deleted commands
add a waitButton command, may not work?  no idea
intakePos fixed
liftVerticalPos fixed
remove commands from starting up upon initializing
upped intake speed :)
added a bunch of soft limits to places too

+++ src/main/java/frc/robot/commands/intake/IntakePosition.java
@@ -14,9 +14,9 @@ public class IntakePosition extends Command {
     
     private double _target;
 
-    public IntakePosition(int target) {
+    public IntakePosition(double target) {
         requires(Intake.getInstance());
-        _target = target; // this should be the degree conversion, from the top
+        _target = target;
     }
 
     @Override
